Move order detail and account global components
