+++ app/account/_components/ResetPasswordForm.tsx
@@ -3,7 +3,7 @@ import { useRef, useState } from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { AiOutlineEye, AiOutlineEyeInvisible } from 'react-icons/ai';
 
-import { Spinner } from '../globals/Spinner';
+import { Spinner } from '~/components/globals/Spinner';
 
 import Button from '~/components/globals/button/Button';
