+++ app/account/addresses/_components/NewAddressModal.tsx
@@ -2,7 +2,7 @@ import { useRouter } from 'next/navigation';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { GrClose } from 'react-icons/gr';
 
-import Modal from '~/components/account/Modal';
+import Modal from 'app/account/_components/Modal';
 import Button from '~/components/globals/button/Button';
 import countriesJSON from '~/data/countries.json';
 import swell from '~/lib/SwellJS';
