+++ app/account/payments/page.tsx
@@ -1,6 +1,8 @@
+import CreatePayment from './_components/CreatePayment';
+
+import PaymentCard from './_components/PaymentsCard';
+
 import AccountLayout from 'app/account/AccountLayout';
-import CreatePayment from 'app/account/payments/_components/CreatePayment';
-import PaymentCard from 'app/account/payments/_components/PaymentsCard';
 
 import { getUserInfo } from '~/lib/SwellGraphQL';
