+++ src/layouts/Layout.tsx
@@ -1,7 +1,7 @@
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-import SkeletonPlaceholder from '~/components/account/SkeletonPlaceholder';
+import SkeletonPlaceholder from 'app/account/_components/SkeletonPlaceholder';
 import Footer from '~/components/globals/Footer';
 import Navbar from '~/components/globals/navbar/Navbar';
